Fix formatting issues in `relvar-core/src/database/constraint_manager.rs`
- Ran `cargo fmt --all` to resolve CI formatting failures.
- Verified changes with `cargo check` and `cargo test`.
